% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa_getOTW_2.R
\name{cpa_getOTW_2}
\alias{cpa_getOTW_2}
\title{Obtain the Optimal Time Window (OTW)}
\usage{
cpa_getOTW_2(data, h2)
}
\arguments{
\item{data}{\code{Numeric} \code{Matrix} of dimension (N_genotype x N_days)
containing the G-BLUEs of a particular trait. The data colnmanes must be
in Date compatible format 'dd-mm-yyyy'.}

\item{h2}{\code{Numeric} \code{vector} of dimension (N_days)
containing the heritability estimates on each day for the trait}
}
\value{
List with the following items:

\item{change_points}{\code{Numeric vector} indicating the location of the
change points on the time series.}

\item{change_points_dates}{\code{Character vector} dates of the time series
change points.}

\item{OTW_data}{\code{List} of \code{Numeric matrices} representing the
genotype trait values within the OTW.}

\item{OTW_data_opt}{\code{Numeric matrix} with the  genotype trait values
within the OTW with the highest heritability.}
}
\description{
Perform Change Point Analysis (CPA) on genotype adjusted means or G-BLUEs
time series of a trait
}
\details{
Entire temporal data set is partitioned into windows based on the differences
in the distribution of trait heritability and phenotypic separability using a
change point analysis procedure defined by Matteson et al. (2014).
}
\examples{

data(SG_PH_data)
SG_PH_data$col_f <- factor(SG_PH_data$col)
SG_PH_data$row_f <- factor(SG_PH_data$row)

SG_PH_data$rep <- factor(SG_PH_data$rep)
SG_PH_data$block <- factor(SG_PH_data$block)

exp_des_data = SG_PH_data[, c("row", "col", "row_f", "col_f","genotype",
                              "rep", "block")]

\dontrun{

op <- SpaTemHTP_proc(exp_des_data, pheno_data = SG_PH_data[, 6:28],
                     out_det = TRUE, miss_imp = TRUE, sp_adj = TRUE,
                     random = ~ rep +  rep:block + row_f + col_f,
                     h2_comp = TRUE, plot = TRUE)
                     
data <- op$G_BLUES

# make sure data colnmanes are dd-mm-yyyy Date format compatible
dates <- substr(colnames(data), 2, nchar(colnames(data)))
dates <- str_replace_all(string = dates, pattern = '\\\\.', replacement = '-')
colnames(data) <- dates

h2 <- op$h2

OTW <- cpa_getOTW_2(data = data, h2 = h2)

}

}
\references{
Matteson, D.S. and James, N.A. (2014). A nonparametric approach for multiple
change point analysis of multivariate data. Journal of the American
Statistical Association, 109(505), pp.334-345.
}
\author{
Soumyashree Kar, Vincent Garin
}
