% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_comp_miss.R
\name{geno_comp_miss}
\alias{geno_comp_miss}
\title{Identify the genotype with missing values}
\usage{
geno_comp_miss(d)
}
\arguments{
\item{d}{\code{data.frame} containing the measured phenotypic values.}
}
\value{
Return:

\code{character} vector with genotype having only missing values.
}
\description{
Identify the genotype with missing values
}
