% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_imp_PMM.R
\name{miss_imp_PMM}
\alias{miss_imp_PMM}
\title{Missing values imputation using PMM}
\usage{
miss_imp_PMM(data, plot = TRUE)
}
\arguments{
\item{data}{\code{data.frame} of dimension (N_genotype * N_replicate) x N_days
containing the measured phenotypic values.}

\item{plot}{\code{logical} value. If plot = TRUE, plot an overview of the
missing values pattern. Default = TRUE.}
}
\value{
Return:

\code{data.frame} with missing values imputed.
}
\description{
Impute the missing values using predictive mean matching
}
\details{
Missing values are imputed sequentially from the first to the
last day.
}
\examples{


data(SG_PH_data)

data <- outliers_det_boxplot(data = SG_PH_data[, 6:28])

\dontrun{

data <- miss_imp_PMM(data = data)

}



}
\references{
Rubin, D. B. (1986). Statistical matching using file concatenation with
adjusted weights and multiple imputations. Journal of Business & Economic
Statistics, 4(1), 87-94.
}
\author{
Soumyashree Kar, Vincent Garin
}
