% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TS.R
\name{plot_TS}
\alias{plot_TS}
\title{Plot G-BLUEs time series data}
\usage{
plot_TS(data_TS, trend = NULL, h2 = NULL, TW_lim = NULL, main = "G-BLUEs TS")
}
\arguments{
\item{data_TS}{\code{numeric} \code{matrix} of dimension N_genotype 
x N_days containing the phenotype information. the different columns
represent the different time points (e.g. days, hours) of measurement.}

\item{trend}{Optional \code{numeric} \code{matrix} of dimension n_days containing the
values of the trend to be ploted on the top of the time series Default = NULL.}

\item{h2}{Optional heritability \code{numeric vector} that will be used to
indicate the day with the highest heritability. Default = NULL.}

\item{TW_lim}{Optional time window limits \code{numeric vector} that will be
used to indicate the time window limit days. Those values can be obtained with
\code{\link{cpa_getOTW_2}}. Default = NULL.}

\item{main}{\code{Character} string representing the title of the graph. Default =  'G-BLUEs TS'.}
}
\value{
Plot of the G-BLUEs time series. The optional trend will be ploted in red.
The heritability will be ploted in dashed blue and time window limits in black.
}
\description{
Plot G-BLUEs time series data.
}
\details{
The function plots the times series data of phenotyping data per genotype
replicate or averaged over genotype replicate using the ggplot2 package.
}
\examples{

data(SG_PH_data)
data <- SG_PH_data
geno_id <- data$genotype

g_means <- matrix(NA, 384, 23)

for(i in 1:23){
  
  g_av <- tapply(X = data[, i + 5], INDEX = geno_id,
                 FUN = function(x) mean(x, na.rm = TRUE))
  g_means[, i] <- g_av
  
}

rownames(g_means) <- names(g_av)

plot_TS(data_TS = g_means, main = 'Raw data genotype means')


}
\seealso{
\code{\link{SpaTemHTP_proc}}
}
\author{
Soumyashree Kar, Vincent Garin
}
