% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TS_raw.R
\name{plot_TS_raw}
\alias{plot_TS_raw}
\title{Plot raw phenotypic time series data}
\usage{
plot_TS_raw(data_TS, geno_id, rep_av = TRUE, main = "TS")
}
\arguments{
\item{data_TS}{\code{numeric} \code{matrix} of dimension (N_genotype x N_rep) 
x N_days containing the phenotype information. the different columns
represent the different time points (e.g. days, hours) of measurement.}

\item{geno_id}{\code{character} vector representing the genotype indicator
eventually replicated a number of time.}

\item{rep_av}{\code{Logical} value indicating if the value should be averaged
over replicates of no. Default = TRUE.}

\item{main}{\code{Character} string representing the title of the graph.
Default =  'TS'.}
}
\value{
Plot of the G-BLUEs time series. The optional trend will be ploted in red.
The heritability will be ploted in dashed blue and time window limits in black.
}
\description{
Plot raw phenotypic time series data
}
\examples{

data(SG_PH_data)
data_TS <- SG_PH_data[, 6:28]
geno_id <- SG_PH_data$genotype

plot_TS_raw(data_TS, geno_id, rep_av = FALSE)

plot_TS_raw(data_TS, geno_id, rep_av = TRUE,
main = 'TS averaged over replicates')


}
\author{
Vincent Garin
}
