\name{SG_PH_data}
\alias{SG_PH_data}
\docType{data}
\title{
  Example high throuput time series data
}
\description{

Example high throuput time series data from a sorghum association panel measured during 23 days for plant height on the LeasyScan platform (Vadez et al. 2015). The association panel was composed of 384 lines laid out as an alpha design with 12 blocks. The experiment was conducted by Milan O. Urban.


}
\usage{data(SG_PH_data)}
\format{
  \code{data.frame}
}
\details{

...

}

\references{

Vadez, V., Kholová, J., Hummel, G., Zhokhavets, U., Gupta, S.K. and Hash, C.T. (2015). LeasyScan: a novel concept combining 3D imaging and lysimetry for high-throughput phenotyping of traits controlling plant water budget. Journal of Experimental Botany, 66(18), pp.5581-5593.

}

\examples{
  data(SG_PH_data)
}

\keyword{datasets}
