% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTemHTP_proc.R
\name{SpaTemHTP_proc}
\alias{SpaTemHTP_proc}
\title{HTP data processing}
\usage{
SpaTemHTP_proc(
  exp_des_data,
  pheno_data,
  out_det = TRUE,
  miss_imp = TRUE,
  sp_adj = TRUE,
  single_mixed_model = FALSE,
  out_p_val = 0.05,
  print_day = TRUE,
  fixed = NULL,
  random = ~row_f + col_f,
  h2_comp = FALSE,
  h2_comp_alt = FALSE,
  plot = TRUE
)
}
\arguments{
\item{exp_des_data}{\code{data.frame} of dimension (N_genotype * N_replicate)
x N_variable containing the experimental design information. It must include:
a) a 'genotype' column representing the line phenotyped;
b) numeric 'row' and 'col' column representing the row and column informaiton,
c) the same row and column information into factor columns named 'row_f' and
'col_f'. Other variables like replicate or block can be introduced to be used
in the spatially adjusted mixed model computation. The user must set those
extra variable in the correct format (generally factor).}

\item{pheno_data}{\code{data.frame} of dimension (N_genotype * N_replicate)
x N_days containing the measured phenotypic values.}

\item{out_det}{\code{Logical} value specifying if outlier detection should
be performed on the phenotypic data. Default = TRUE.}

\item{miss_imp}{\code{Logical} value specifying if missing value imputation
should be performed on the phenotypic data. Default = TRUE.}

\item{sp_adj}{\code{Logical} value specifying if a mixed model with spatial
adjustment (SpATS model) should be used to calculate the genotype BLUEs.
Default = TRUE.}

\item{single_mixed_model}{\code{Logical} value indicating if a 'single-step'
mixed model should be calculated. See Details for more explanations.
Default = FALSE.}

\item{out_p_val}{\code{Numeric} value indicating the signficance threshold
for outliers detection. Default = 0.05.}

\item{print_day}{\code{Logical} value indicating if the day progression should
be printed. Default = TRUE.}

\item{fixed}{Optional right hand formula object specifying the fixed effects
of the SpATS model. Default = NULL.}

\item{random}{Optional right hand formula object specifying the random effects
of the SpATS model. Default = ~ row_f + col_f.}

\item{h2_comp}{\code{Logical} value indicating if the should calculate the daily
genotypic heritability. Default = FALSE.}

\item{h2_comp_alt}{\code{Logical} value indicating if the should calculate the daily
genotypic heritability. Default = FALSE.}

\item{plot}{\code{Logical} value specifying if a time series plot of the
G-BLUEs should be produced. Default = TRUE.}
}
\value{
Return:

If sp_adj = FALSE, the code to calculate the genotype BLUEs using a mixed
model without spatial adjustment is not available, so we return
the matrix of data after eventual processing operation (outlier detection,
missing value imputation). Those data can be used with another software
(e.g. Genstat) to calculate the genotype BLUEs without spatial adjustment.

If sp_adj = TRUE, list containing the following objects

\item{G_BLUES}{\code{Matrix} of the genotype BLUEs with the genotype in row
and the day (or measurement time) in column.}

\item{G_BLUES_stdev}{\code{Matrix} of the genotype BLUEs standard deviation
with the genotype in row and the day (or measurement time) in column.}

\item{h2}{\code{Vector} with the daily genotypic heritability.}
}
\description{
Calculate sequentially the genotype adjusted means (genotype best unbiased
linear estimates, G-BLUEs).
}
\details{
The function iterates over the different time points (e.g days) of the
experiment and calculate the G-BLUEs according to the following model:
pheno = Int + row(random) + col(random) + genotype(fixed) + f(row, col) + e,
where f(row, col) represent the spatial surface modeled using a
2-D P-spline approach as proposed by Rodgriguez-Alvarez et al. (2018).

The user can run a different model using the fixed and random arguments
that specifies the fixed and random part of the mixed model used to
calculate the genotypes BLUEs. The spatially adjusted model are fitted
using function from the SpATS package (Rodgriguez-Alvarez et al., 2018).

If \code{single_mixed_model = TRUE}, the function calculates a single-step mixed
model where outliers are iteratively removed based on the model residuals
and the missing values imputed during the estimation procedure. The outliers
are detected using a Grubb Test (p<0.05, default).

If \code{h2_comp = TRUE}, the function calculate the heritability from the
computed SpATS model. If \code{h2_comp_alt = TRUE}, if the heritability
computation fail with the SpATS model, the function calculates the
heritability using a classicle linear mixed model (pheno = rep (F) + 
row (R) + col (R) + genotype (R) + e). The formula used in the following
h2 = Sgeno / (Sgeno + (Se/n_rep)).
}
\examples{

data(SG_PH_data)

SG_PH_data$col_f <- factor(SG_PH_data$col)
SG_PH_data$row_f <- factor(SG_PH_data$row)

SG_PH_data$rep <- factor(SG_PH_data$rep)
SG_PH_data$block <- factor(SG_PH_data$block)

exp_des_data = SG_PH_data[, c("row", "col", "row_f", "col_f","genotype",
"rep", "block")]

\dontrun{

G_BLUEs <- SpaTemHTP_proc(exp_des_data, pheno_data = SG_PH_data[, 6:28],
                          out_det = TRUE, miss_imp = TRUE, sp_adj = TRUE,
                          random = ~ rep +  rep:block + row_f + col_f,
                          plot = TRUE)

}

}
\references{
Maria Xose Rodriguez-Alvarez, Martin P. Boer, Fred A. van Eeuwijk, Paul
H.C. Eilers (2018). Correcting for spatial heterogeneity in plant breeding
experiments with P-splines. Spatial Statistics 23 52 - 71
URL https://doi.org/10.1016/j.spasta.2017.10.003
}
\author{
Soumyashree Kar, Vincent Garin
}
