% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_trend.R
\name{TS_trend}
\alias{TS_trend}
\title{Time series trend modelling}
\usage{
TS_trend(data, linear = TRUE, logistic = TRUE)
}
\arguments{
\item{data}{\code{numeric} \code{matrix} of dimension N_genotype 
x N_days containing the G-BLUEs time series. Such an object can be obtained
with the function \code{\link{SpaTemHTP_proc}}.}

\item{linear}{\code{Logical} value specifying if the linear trend should be
fitted. Default = TRUE.}

\item{logistic}{\code{Logical} value specifying if the logistic trend should be
fitted. Default = TRUE.}
}
\value{
Return:

\item{lin_res}{\code{list} containing the parameters of the G-BLUEs TS
linear trends.}

\item{lin_R2}{\code{Vector} of r squared goodness of fit statistic
for each G-BLUEs TS modeled with the linear trend.}

\item{log_res}{\code{list} containing the parameters of the G-BLUEs TS
linear trends.}

\item{log_R2}{\code{Vector} of r squared goodness of fit statistic
for each G-BLUEs TS modeled with the logistic trend.}

\item{log_val}{\code{matrix} of predicted values according to the
fitted logistic trend for each G-BLUEs TS.}
}
\description{
Fit linear and logistic trends to G-BLUEs time series.
}
\details{
The function fit a linear and/ or a logistic function to the genotype
best linear unbiased estimates time series obtained with the function
\code{\link{SpaTemHTP_proc}}. The logistic function is fitted using
the function drm from package drc.
}
\examples{

data(SG_PH_data)

SG_PH_data$col_f <- factor(SG_PH_data$col)
SG_PH_data$row_f <- factor(SG_PH_data$row)

SG_PH_data$rep <- factor(SG_PH_data$rep)
SG_PH_data$block <- factor(SG_PH_data$block)

exp_des_data = SG_PH_data[, c("row", "col", "row_f", "col_f","genotype",
"rep", "block")]

\dontrun{

G_BLUEs <- SpaTemHTP_proc(exp_des_data, pheno_data = SG_PH_data[, 6:8],
                          out_det = TRUE, miss_imp = TRUE, sp_adj = TRUE,
                          random = ~ rep +  rep:block + row_f + col_f,
                          plot = TRUE)
                          
G_TS_trend <- TS_trend(G_BLUEs)

}

}
\references{
Ritz, C., Baty, F., Streibig, J. C., Gerhard, D. (2015) Dose-Response
Analysis Using R PLOS ONE, 10(12), e0146021
}
\seealso{
\code{\link{SpaTemHTP_proc}}
}
\author{
Soumyashree Kar, Vincent Garin
}
