% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TS.R
\name{plot_TS}
\alias{plot_TS}
\title{Plot G-BLUEs time series data}
\usage{
plot_TS(data_TS, trend = NULL, h2 = NULL, TW_lim = NULL, main = "G-BLUEs TS")
}
\arguments{
\item{data_TS}{\code{numeric} \code{matrix} of dimension N_genotype 
x N_days containing the G-BLUEs time series. Such an object can be obtained
with the function \code{\link{SpaTemHTP_proc}}.}

\item{trend}{Optional \code{numeric} \code{matrix} of dimension n_days containing the
values of the trend to be ploted on the top of the time series Default = NULL.}

\item{h2}{Optional heritability \code{numeric vector} that will be used to
indicate the day with the highest heritability. Default = NULL.}

\item{TW_lim}{Optional time window limits \code{numeric vector} that will be
used to indicate the time window limit days. Those values can be obtained with
\code{\link{cpa_getOTW_2}}. Default = NULL.}

\item{main}{\code{Character} string representing the title of the graph. Default =  'G-BLUEs TS'.}
}
\value{
Plot of the G-BLUEs time series. The optional trend will be ploted in red.
The heritability will be ploted in dashed blue and time window limits in black.
}
\description{
Plot G-BLUEs time series data.
}
\details{
The function plots (G-BLUEs) times series data organized by genotype using the
ggplot2 package. Each line represent a specifice genotype or individual.
Using argument 'trend', it is possible to provided an averaged trend over the genotypes
or any other trend that will be plot on the top of the time series.
}
\examples{

data(SG_PH_data)
data <- SG_PH_data
geno_id <- data$genotype

g_means <- matrix(NA, 384, 23)

for(i in 1:23){
  
  g_av <- tapply(X = data[, i + 5], INDEX = geno_id,
                 FUN = function(x) mean(x, na.rm = TRUE))
  g_means[, i] <- g_av
  
}

rownames(g_means) <- names(g_av)

plot_TS(data_TS = g_means, main = 'Raw data genotype means')


}
\seealso{
\code{\link{SpaTemHTP_proc}}
}
\author{
Soumyashree Kar, Vincent Garin
}
